+++ examples/src/main.cpp
@@ -15,6 +15,7 @@
 #include "scenes/RectangleScene.h"
 #include "scenes/RectangleCameraScene.h"
 #include "scenes/CubeScene.h"
+#include "scenes/CubeTextureScene.h"
 
 // NOTE: Initialize pointer to zero so that it can be initialized in first call to getInstance
 // Because main is the first cpp file to compile and read we instantiate SceneManager here
@@ -32,8 +33,10 @@ int main(int argc, char *args[]) {
     game.getSceneManager().addScene("rectangle", std::make_shared<RectangleScene>());
     game.getSceneManager().addScene("rectangle_camera", std::make_shared<RectangleCameraScene>());
     game.getSceneManager().addScene("cube", std::make_shared<CubeScene>());
+    game.getSceneManager().addScene("cube_texture", std::make_shared<CubeTextureScene>());
 
-    game.getSceneManager().setActiveScene("cube");
+
+    game.getSceneManager().setActiveScene("cube_texture");
 
     game.run(); // run the game loop